Fix for bug #876: Windows 7/8/10 - Pinning Jabref to the taskbar now works
Details:
Previously, JabRef couldn't be pinned to the taskbar ([bug report](https://sourceforge.net/p/jabref/bugs/876/)). The reason is that, by wrapping JabRef with Launch4j, Windows thinks only Java is running and does not recognize JabRef as something separate.

Based on a [stackoverflow question](http://stackoverflow.com/questions/9330209/whats-the-best-way-to-start-java-applications-on-windows-7) the fix proceeds in two stages:
* During the setup, we create a (dummy) shortcut and set the [Application User Model Id](https://msdn.microsoft.com/en-us/library/windows/desktop/dd378459%28v=vs.85%29.aspx). For this, we use the [WinShell plugin](http://nsis.sourceforge.net/WinShell_plug-in) for NSIS.
* At program start, we notify Windows that this JabRef instance has the AppUserModelId specified in the setup. In order to do this, we invoke native commands using the Java Native Access (JNA) library.