Implementation of the about dialog in JavaFX with afterburnerfx
The old dialog and html layout file have been removed.
The layout of the dialog panel is defined in the AboutDialog.fxml file
located in the resources with its stylesheet. The AboutDialogViewModel
class takes care of the text initialization of any components in the
layout. It also implements any methodes called from any component of the
dialog panel. The dialog is build and shown in the AboutDialogView
class.