Addin two tests and fixing one bug in 'BracketedExpressionExpander'.
To avoid 'uninitialised toolkit' exceptions from FX, had to resort to an
ugly solution -- creating an unused JFXPanel instance in the text. For
the command line driver even this does not work, since the programs does
not exit when JFXPanel is created (and I do not know at the moment how
to destroy it).