Fix timestamps not being updated for changed entries
This is a fix for the issue #2810.

The functionality has been moved from the EntryEditor to the BasePanel
class so that it can listen for EntryChangedEvents on the whole
database instead of specific EntryEditor objects. This allows us to
also update the timestamps of entries that were changed by other
sources (like cleanup operations).

The inner class UpdateTimestampListener is defined as "public static"
so that it can be used in the test cases for the implemented
functionality.