Fix for issue #3381 (#3443)
* the SearchDescribers now return a TextFlow

* moved the search.rules.describer package from logic to gui package
  because the SearchDescriber now creates TextFlow Elements

* moved the tests accordingly

* modified creation of Tooltip in GlobalSearchBar to display the
  TextFlow created by the SearchDescribers

* SearchQuery now stores a TextFlow as description

* Added a stub for compatibility in GroupDialog because it uses the
  SearchQuery to get the description and needs an HTML string for
  its Swing Tooltip

* Styled the SearchBar Tooltip for readability

It uses an own style class to not change any other styled elements.
Styling was choosen to look like the other (swing) tooltips.

* refactor for better readability

* removed description member of SearchQuery class

because the description of a SearchQuery is a gui element, it is not
stored any more in this class. Everywhere the description is needed it
is generated

* add method to create HTML string from javafx Text

* create HTML string from description for compat.

* fixed formatting for style

* created tests for TextUtil class

* modified the SearchDescriber Implementations tests

now they work properly again and check the created Texts inside
the description TextFlow

* split up testcases into single methods

* using old localized strings again

these old localized strings containing html tags are broken up and
converted to Text objects to use in javafx TextFlow

* added and modified tests

* added changelog entry

* changed syntax for getting SearchDescriber

now a whole SearchQuery object is sufficient to get a corresponding
SearchDescriber

* refactored for better style

* javafx tooltips styled per css

also refactored the GlobalSearchBar tooltip to use this new styling

* modified and restructured tests

It was necessary to modify the tests of the SearchDescriber
Implementations and the TooltipTextUtil class because of the new css
styling for tooltips. Also refactored for better style