Fix the download encoding - doesn't fit library encoding though
The problem is that the matching of strings (find abstract, ... in the HTML)
is done against strings in the source code - therefore they are encoded with
the source code encoding.
The downloaded HTML however should be encoded with whatever the user selected
for her library. But then, the matching fails.
So this needs further investigation.