Fix the IACR fetcher for entries created before year 2000
The entries before year 2000 use a slightly different format which e.g. doesn't
include a version, also the date format is different.

With this commit, we also throw an error if the user tries to fetch an entry
for a withdrawn paper. This is meant as a warning to the user, she might still
add the entry manually to her database.
This will be especially useful once a "search by title" or something similar
gets implemented.