Fetcher for IACR eprints (#3473)
* Add an initial WIP version of an IACR eprint fetcher

Note: This is WIP, the fetcher kind of worked in a few ad-hoc tests,
but it certainly isn't ready for production yet!

* Refactor IACR fetcher and improve error handling

* Localize error messages in IACR fetcher

* Refactoring of IACR fetcher

* More refactoring of IACR fetcher

By moving all the "what is the content of field xy?" logic to methods,
the setAdditionalFields method gives a good overview over which fields
are set.

* Fix the download encoding - doesn't fit library encoding though

The problem is that the matching of strings (find abstract, ... in the HTML)
is done against strings in the source code - therefore they are encoded with
the source code encoding.
The downloaded HTML however should be encoded with whatever the user selected
for her library. But then, the matching fails.
So this needs further investigation.

* Add tests for IACR fetcher

* Make checkstyle happy

* Add IACR fetcher to changelog

* Fix the encoding of entries retrieved via IACR fetcher

* Migrate IACR fetcher tests to junit5

* Use enum constant for IACR fetcher import encoding

* Fix a bug in the IACR fetcher tests

* Migrate IACR fetcher to java8 date classes

* Remove abstracts from IACR fetcher tests

As pointed out by @Siedlerchr in #3473, the abstracts might
be a copyright problem.

* Fix the IACR fetcher for entries created before year 2000

The entries before year 2000 use a slightly different format which e.g. doesn't
include a version, also the date format is different.

With this commit, we also throw an error if the user tries to fetch an entry
for a withdrawn paper. This is meant as a warning to the user, she might still
add the entry manually to her database.
This will be especially useful once a "search by title" or something similar
gets implemented.

* Make the localization strings in the IACR fetcher reusable

Also: Remove some of them completly, replacing them with slightly
different existing ones.

* Disable a long running test for the IACR fetcher

* Fix codacy issues in the IACR eprint fetcher

codacy complained about reassigning a method parameter
and about the visibility of a test method.

* Merge localization files with master

* Fix duplicate keys in localization files