Add java property to make location of Main.css customizable
The "live reloading" has now to be turned on by defining -Djabref.main.css
Additionally, the location of the Main.css file can be given as parameter to
this property. This makes it way more convenient when working on the code, because
usually the location of the Main.css is in the build folder. Now, you can
use the Main.css from the source directly.