Change some icons for consistent visual weight and additional adaptions
In the main toolbar, I changed the "New File" icon because it was an
outlined version while Open and Save is a filled version. For the Open
icon I would like to have the "folder-open", but that is currently not included
in our version.

I changed the Add Entry icon to a filled version as well, since it better fits
the trash. Unfortunately, there is no filled version of Copy and Paste and the
positively filled version of GitHub looks just awful.

I changed the Up, Down, and Close icon for the sidepanel and the entry editor
to versions that look visually more similar in size. Especially, the up and down
was too small.

After correspondence with a friend, I changed the default list item before the
groups into something not so heavy. The small triangle with its vertical side on the
left gives additionally a better impression of the indentation which is the
key here.

The Add Group icon was changed into a heavier circled version. It
now matches the style of the close button of the sidepanel and entry editor.

All icons of "sub toolbars" in the entry editor and the side panel have now
an equal size of 16px.

The preferred height of the side-panel header and the header of the tab-pane
was adjusted so that they match and there is no jump between main-table and
side-panel anymore. There might be a better way to do this.

Finally, the color of the overflow buttons for toolbar and tab-pane was
adjusted.