Improve small details
Base.css:
- Introduce explicit colors for hover/select of table. This needs to be
adjust for dark theme (-jr-hover-color, -jr-select-color)
- Additional color for hovered menu text. No influence on light theme but
might be needed in the dark theme -jr-menu-foreground-active
- A slightly darker gray for main text
- reenabled hover effect for icons
- reenabled theme color for tab labels. If required for dark theme lets
introduce a separate color for it
- fixed tab close button
- in light theme we need a visible border for the split pane. Otherwise e.g.
main table and preview have no visible border.
- removed that a selected entry in table has a different color when the maintable
is not focused.
- fixed hover highlighting for menu

GroupTree.css
- the bottom border is not required, but with the search bar in white,
we need a visible border to the group tree itself

ErrorConsole.css
- fixed the style

EntryEditor.css
- Same color for code area then the rest

Entry Editor code area uses the same font color as the rest of the