Fix colors and button effects
Buttons now have a simply hover effect as suggested by Tobi. To make this
work consistently, I *did not* use a color for the effect. Instead,
simple black (or white for the dark-scheme) is used with a high value of
alpha. This works independent of the underlying background by just darkening
the area a bit.

I removed the second hover color and replaced it by alpha darkening as well.

Other small fixes include the toolbar context menu and adjustments to
group view and entry editor.

Finally, I tried to bring the changes to the dark version.