Convert CiteSeerX fetcher to new infrastructure
The old implementation first extracted the detail pages for each article matched by the query and then went to every detail page to extract the bibliographic information. Instead, we now parse the COinS information that is already contained in the main result page and thereby reduce the number of requests to 1 per query.