Post change notifications on JavaFX (#4871)
* Post change notifications on JavaFX

Fixes #4817. The problem is that changes in other threads produce a "Not on FX application thread" exception. This is fixed by adding a wrapper around the list of entries. The wrapper only posts changes in the correct thread, without making a copy of the underlying list - so performance shouldn't be affected.

* fix compile error

* fix checkstyle