Fix cursor jump
Fixes #5084. The entry editor was very aggressive and stole the focus from the search bar or from the main table. This lead to problems like #5084 where users overwrite the contents of the entry while doing a search. Moreover, it made it impossible to navigate between entries using the arrow keys in the main table. Both of these issues are fixed now.