[feat] Implement mockup of ShortScience integration
This commit adds a simple integration of ShortScience.org
in the form of an additional context menu action when an
entry is selected. This action opens the user's web browser
to the search page of ShortScience with the title of the
paper entered. Essentially, this is our MVP for the
integration towards ShortScience bar some testing.

[Issue: #3]