Changed to Task<?> in the Tasklist.
This makes the view work with download tasks for example. Most other
tasks are still shown without title, message (because none are set) and
progress.
Also, there are a lot of tasks somehow.

The progress indicator in the main view still does not work as I can't
get the bindings to work.