Converted list of Properties to tasks for listbind
When using ObjectProperties in the list of tasks, we can use EasyBind to
convert the list into a list of tasks which in turn can be bound to the
list of tasks in the view.

With this, the basic functionality works. There is a progress indicator
in the toolbar that shows the average progress of all running background
tasks. It is indeterminate if any task has indeterminate progress and
shows 100% if no tasks are running.

Clicking it opens an overview of all running tasks and their progress.

Currently, there are many tasks running all the time. The only tasks
that were adapted for this to be pretty are the download tasks, and they
are also still missing an icon.