More informative (and working) quit dialog
The dialog that is shown when the user quits JabRef during ongoing
background tasks now shows which background tasks are still running.

When all of them complete or the user chooses to quit anyway, JabRef
quits.

The user can also cancel the dialog. In that case, the dialog closes and
JobRef remains running.