Change "duplicates" window to recall size/position
The "Possible duplicate entries" window now remembers its size and position throughout a session ([#6582](https://github.com/JabRef/jabref/issues/6582)). This was done by adding a DialogWindowState class to the gui utils. This class stores the size and position of the dialog window. The StateManager class was modified to store this object. It is retrieved whenever a new DuplicateResolverDialog object is created to retrieve the previous size and position and is updated whenever the user proceeds through the dialog options. The DialogWindowState class can be re-used if this feature is needed for other dialog windows.

This is my first ever pull request so please let me know what you think or whether I need to re-tool my approach to be more in line with the structure of the rest of the code! I did not include any screenshots as this felt like a sort of awkward fix to try and take screenshots of.