Refactor DialogWindowState to be immutable
DialogWindowState is now immutable, and is stored in a map in StateManager specifically for DialogWindowStates. I could not manage to inject the StateManager directly into the DuplicateResolverDialog class, I assume because it does not use a view model in the same way that other dialog/views do throughout JabRef, so instead I just passed it in as a parameter and modified the other files that call DuplicateResolverDialog to reflect this.

The question remains as to whether the StateManager is the best place to store this sort of information. As it's only stored for the session, it works fine as long as the user doesn't mind having to reset the size of the DuplicateResolverDialog window when they reload the program.