Use name of class instead of magic string
I updated the set/getDialogWindowState methods to use the name of the class instead of magic strings. I looked into updating the preferences to store the state but I'm sort of confused by how the preferences are currently managed, as there are plenty of classes that still use Globals to access the preferences.

It seems like either the StateManager approach or the preferences approach should be used, but not both at the same time. If I pass the preferences to the DuplicateResolverDialog class, I could save the position of the window right there, but this makes the DialogWindowState class basically pointless. Otherwise, I could save the state into the preferences at some juncture but I'm not sure where the best place to do this is, especially if I want to avoid using globals.