Fix locating jabref in jabrefHost.py
shutil.which() returns an str and not a pathlib.Path. Therefore, we need
to check if which() actually returns something and if so, we should
convert it to a pathlib.Path such that JABREF_PATH is always of the same
type.

Also, try locating JabRef in addition to jabref in the user's PATH.