Fix issue 6936.
Issue #6936 is about the source tab (BibTeX source) loosing the ability
to be editable (https://github.com/JabRef/jabref/issues/6936#issuecomment-703110024).

This patch re-enables codeArea as editable if the source code is correct. Although
the source tab cannot be used by the user to fix any syntax error in BibTeX fields,
such error can be fixed by editing the field, which will trigger SourceTab's listener
and call the updateCodeArea() method, which will enable codeArea if the source of the
 BibTeX entry is syntactically correct.