compare content in GroupDiffTest.java
Made private constructor package-private to be able to create a GroupDiff instance for testing.

The test fails when comparing the Optional<GroupDiff> objects directly due to the different object IDs, despite the fact that their fields should be identical.
Therefore, I decided to compare the fields of the objects to check if they are equal.