Initial draf of abbreviation search feature
Implemented a string filter for the journal abbreviations tab.
Not working. Performing a search currently cleares a list of filtered
abbreviations and fills it back up with abbreviations that contain the
search string. Right now the clearing not only cleares the filtered list
but also the list of all abbreviations and I don't unterstand why.

Issue #7751