Prioritize non-bruteforce importers that
When importing, try importers that can tell if they are suitable for a
certain file format or not.
Some importers only check if a file is present, not if it in the correct
format (isRecognizedFormat is always true if an existing file is given).
They are used last.

The List of importers now reflects that prioritization. It is not sorted
by importer names anymore.
The getter-methods getImportFormats and getImportFormatList still sort
the List by name for the View.