Added new Fetcher that mimics ArXiv fetcher, but also gets new fields from ArXiv-issued DOI
The new Fetcher, 'ArXivWithDoi', implements the same interfaces as the 'ArXiv' fetcher and, for the most part, rely on calling the latter for complying with these interfaces. The actual innovation is that all BibEntries returned from 'ArXiv' are replaced by a merged version between it and a BibEntry from the DoiFetcher (with ArXiv's fields getting the priority)

Additionally, two other changes were made:

- Method 'merge()' to BibEntry, so that two BibEntries can be merged (with priority given to the current 'BibEntry' object)
- FOR NOW, all references to the use of the 'ArXiv' class in other parts of the database (```WebFetchers```, ```ImportHandler``` and ```CompositeIdFetcher```) were replaced with the new 'ArXivWithDoi' class, so that all UI path lead to the use of this new version. Future commits may make its use optional with feature flags

A small number of manual tests were made. More to come. Next commits should also integrate automated tests.