Added API Rate Limiting for async calls to DOI API (DOI Content Negociation)
As a previous worry about using multithreading to make the extended
ArXiv fetching process, API throlling SHOULD be mostly avoided, with the
integration of a limiter for agencies "DataCite" (used by ArXiv-issued
DOIs) and "Crossref". These two agencies define the maximum rate in which
applications can perform API requests via DOI Content Negotiation (mEDRA
does not seem to explicitly, so it's not being considered for now).

For that to be possible, other changes had to be made in URLDownload for
reduing connection overhead.