Add the 'Review backup' button to the dialog and do some experiments
I was attempting to handle the dialog actions within the dialog class, but encountered some roadblocks. For example, when the user clicked restore from backup, I wanted to run the restore backup code without blocking the FX thread, so I wrapped it in a BackgroundTask#wrap. However, the database loading task in OpenDatabaseAction#loadDatabase will continue because the dialog is closed once the action is clicked, and the JVM continues running the code at BackupUIManager#showRestoreBackupDialog.