Update MacOS jabrefHost.py to find local installs (#9487)
Following issue [9474](https://github.com/JabRef/jabref/issues/9474).

The previous MacOS hardcoded path fails if JabRef is installed to a users local applications directory rather than the system one (e.g. if the user doesn't have admin permissions). On my version of MacOS (12.6.1) this change works correctly for local installs as well. The Mac app bundle is structured differently so an alternate path is still required.

If the difference in `which` behaviour is still a problem (maybe on older versions?) the original hardcoded fallback could also be included as a final option. If that is still necessary the explanatory comment should be updated too.

As mentioned on the issue thread the python script path in the browser extension JSON would still need to be updated for local installs, so I will submit a separate PR to the docs with notes on the process. If this script can't be updated I can expand the docs update to detail the hardcoded path change required to make the browser extension work for local app installs.