Grammar fix and rewording in README.md
Small grammar fix and rewording, replacing "in one of the many thousand built-in citation styles" with "using one of thousands of built in citation styles".
I reworded it a bit, so if people don't like the rewording that's more than fair. 
If the original wording is preferred, then "in one of the many thousands of built in citation styles" would maintain the wording whilst fixing the grammar.