Fix order of string in editor (#10083)
* Fix order of string in editor

The string defined in a BibTeX file were shown in random order in the
respective GUI editor. While the code already sorts them, by storing the
items into a set, it destroys the order again. Fix this behaviour by
storing them as a list, which retains the order.

* Apply suggestions from code review

* Resort the strings table for new entry

Adding a new entry should also preserve the sorting of the strings
table, thus we need to resort the entries and take care that the focus
in the table is still on the newly-created entry.

* Add a test for the resorting

* Cleanup code

* Further cleanup for better reuse