################################################## # Write a title summarizing what this commit does. # Start with an uppercase imperative verb, such as # Add, Drop, Fix, Refactor, Bump; see ideas below. # Think of your title as akin to an email subject, # so you don't need to end with a sentence period. # Use 50 char maximum, which is this line's width. ################################################## Add 'Get New Themes!' Button for Web Page Integration
########################################################################
# Why is this change happening?
# Describe the purpose, such as a goal, or use case, or user story, etc.
# For every line, use 72 char maximum width, which is this line's width.
########################################################################
The change was made at the request of a user who wanted the option to discover new themes for the application more easily and conveniently, such as through a button within the application that redirects them to a web page in their default browser.

########################################################################
# How is this change happening?
# Describe any relevant algorithms, protocols, implementation spec, etc.
# For every line, use 72 char maximum width, which is this line's width.
########################################################################

The change was implemented through a button added in Preferences > General, located just below the option for selecting new themes via a local directory. Modifications were made to the 'GeneralTab.fxml' and 'GeneralTab.java' files, through which the button was created and the method to execute its function was implemented, respectively