Add 'Get New Themes!' Button for Web Page Integration
The change was made at the request of a user who wanted the option to discover new themes for the application more easily and conveniently, such as through a button within the application that redirects them to a web page in their default browser.

The change was implemented through a button added in Preferences > General, located just below the option for selecting new themes via a local directory. Modifications were made to the 'GeneralTab.fxml' and 'GeneralTab.java' files, through which the button was created and the method to execute its function was implemented, respectively

Some requested changes were resolved.