Fix : If a change happens in an old directory, reload is no longer triggered (JabRef#10585)
I created a Map<Path, FileUpdateListener> so that when a Listener is added to a path, you can recover it if you latter when to remove it. Before, the call to fileUpdateMonitor.removeListener didn’t actually removed the listener because we passed a FileUpdateListener that didn’t match the expected one.

Moreover I created the function addListenerForDirectory, which internally calls addListenerForFile(directory, listener), because in this case
A File and a Directory can be treated the same way. The only purpose of this function is for clarity, it didn’t really make sense to call addListenerForFile with a path representing a directory and not a file as parameter.